+++ docs/source/changelog.rst
@@ -38,6 +38,7 @@ Fixed
 - Setting the ``lmodern`` keyword argument of `.Document` to false will not
   cause invalid LaTeX code anymore.
 - `.Quantity` now correctly splits prefix and unit into seperate commands.
+- `.Quantity` can now handle Celsius.
 
 
 1.0.0_ - `docs <../v1.0.0/>`__ - 2015-11-25